requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cloudinary==1.25.0
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.4
whitenoise==5.2.0
